+++ CHANGELOG.md
@@ -9,6 +9,10 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 
 ## Unreleased - xxxx-xx-xx
 
+### New Features 
+
+- The `Enum` derive now supports fallback variants
+
 ## v3.0.0-beta.3 - 2023-05-13
 
 ### New Features